added dependencies and components
In the app.js, called api to retrieve data, handled change and updated state. In the bookshelf.js, displayed the book in the right shelf through props, and passed the book id to the selected event and displayed the title. In the booklist.js filtered the books depending on shelf to get the book on shelf from app and then displayed the three differenst shelves in main pages with its current books.